refactor(http): extract HTTP utilities into http/ directory
Spec 2.1 complete: Extract HTTP utilities from provider-utils.ts
and request-utils.ts into focused modules.

Create http/ directory with:
- fetch-utils.ts: fetchWithResponseTimeout
- url-utils.ts: joinUrl
- index.ts: public exports

Update imports in:
- ollama-context.ts, provider-catalog.ts
- provider-strategy/fallback/legacy.ts, local.ts, shared.ts
- federation/federated-fallback.ts, federation-helpers.ts
- routes/embeddings.ts

Remove duplicate joinUrl definition from provider-strategy/shared.ts.
Remove functions from original files.

All 579 tests pass.
