+++ src/lib/federation/federation-helpers.ts
@@ -1,4 +1,4 @@
-import { fetchWithResponseTimeout } from "../provider-utils.js";
+import { fetchWithResponseTimeout } from "../http/index.js";
 import { readSingleHeader } from "../request-utils.js";
 
 const FEDERATION_HOP_HEADER = "x-open-hax-federation-hop";
