+++ src/lib/provider-catalog.ts
@@ -2,9 +2,9 @@ import type { ProxyConfig } from "./config.js";
 import type { KeyPool, ProviderCredential } from "./key-pool.js";
 import type { ProviderRoute, ResolvedModelCatalog } from "./provider-routing.js";
 import { parseModelIdsFromCatalogPayload, buildLargestModelAliases, dedupeModelIds } from "./provider-routing.js";
-import { fetchWithResponseTimeout } from "./provider-utils.js";
+import { fetchWithResponseTimeout } from "./http/index.js";
 import { loadDeclaredModels, loadModelPreferences, type ModelPreferences } from "./models.js";
-import { joinUrl } from "./request-utils.js";
+import { joinUrl } from "./http/index.js";
 
 const DEFAULT_CATALOG_ROUTE_TIMEOUT_MS = 15_000;
