+++ src/lib/provider-strategy/local.ts
@@ -2,7 +2,8 @@ import type { FastifyReply } from "fastify";
 
 import type { RequestLogStore } from "../request-log-store.js";
 import { buildForwardHeaders, buildUpstreamHeaders } from "../proxy.js";
-import { fetchWithResponseTimeout, sendOpenAiError, toErrorMessage } from "../provider-utils.js";
+import { fetchWithResponseTimeout } from "../http/index.js";
+import { sendOpenAiError, toErrorMessage } from "../provider-utils.js";
 import { getTelemetry } from "../telemetry/otel.js";
 import {
   joinUrl,
