+++ src/lib/provider-utils.ts
@@ -534,26 +534,6 @@ export async function responseIndicatesQuotaError(response: Response): Promise<b
   }
 }
 
-export async function fetchWithResponseTimeout(url: string, init: RequestInit, timeoutMs: number): Promise<Response> {
-  const controller = new AbortController();
-  const timeoutHandle = setTimeout(() => {
-    controller.abort(new DOMException("The operation was aborted due to timeout", "TimeoutError"));
-  }, timeoutMs);
-
-  const mergedSignal = init.signal
-    ? AbortSignal.any([init.signal, controller.signal])
-    : controller.signal;
-
-  try {
-    return await fetch(url, {
-      ...init,
-      signal: mergedSignal
-    });
-  } finally {
-    clearTimeout(timeoutHandle);
-  }
-}
-
 export function sendOpenAiError(
   reply: FastifyReply,
   statusCode: number,
